+++ simple_live_app/ohos/hvigorfile.ts
@@ -1,6 +1,8 @@
-import { appTasks } from '@ohos/hvigor-ohos-plugin';
+import path from "path";
+import { appTasks } from "@ohos/hvigor-ohos-plugin";
+import { flutterHvigorPlugin } from "flutter-hvigor-plugin";
 
 export default {
   system: appTasks,
-  plugins: []
+  plugins: [flutterHvigorPlugin(path.dirname(__dirname))],
 };